==> notebooks/11_ClinPheWASRes.Rmd <==
---
title: An R Markdown document converted from "11_ClinPheWASRes.ipynb"
output: html_document
---

# Results of PheWAS - Clinical phenotypes

```{r}
library(readr)
library(dplyr, warn.conflicts = FALSE)
library(tidyr, warn.conflicts = FALSE)
library(purrr)
```

```{r}
clinphewasres <- read_tsv("../data/clinphewasres.tsv", show_col_types = FALSE)
head(clinphewasres)
```

```{r}
clinphewasres <- clinphewasres %>%
    filter(!is.na(eaf)) %>%
    transmute(rsid_proxy = rsid, ea_exp = ea, nea_exp = nea, eaf_exp = eaf, 
              id_exp = id, name_exp = trait, beta_exp = beta, se_exp = se)
head(clinphewasres)
```

```{r}
proxiestab <- readr::read_tsv("../data/proxiestab.tsv", show_col_types = FALSE)
head(proxiestab)
```

```{r}
clinphewasres <- inner_join(proxiestab, clinphewasres, by = "rsid_proxy", relationship = "many-to-many")
head(clinphewasres)
```

```{r}
clinphewasres <- clinphewasres %>%
    mutate(
        across(
            c(ea_exp, nea_exp),
            ~stringr::str_replace_all(.x, c("A" = "t", "T" = "a", "C" = "g", "G" = "c")),
            .names = "{.col}_flip"
        ),
        across(
            c(ea_exp_flip, nea_exp_flip),
            ~toupper(.x)
        ),
        harmon1 = case_when(ea_proxy == ea_exp & nea_proxy == nea_exp ~ 1,
                            ea_proxy == nea_exp & nea_proxy == ea_exp ~ -1,
                            TRUE ~ NaN),
        harmon2 = case_when(ea_proxy == ea_exp_flip & nea_proxy == nea_exp_flip ~ 1,
                            ea_proxy == nea_exp_flip & nea_proxy == ea_exp_flip ~ -1,
                            TRUE ~ NaN),
        harmon = coalesce(harmon1, harmon2),
        beta_exp = beta_exp * harmon,
        eaf_exp = ifelse(harmon == 1, eaf_exp, 1 - eaf_exp)
    ) %>%
    filter(!is.na(harmon)) %>%
    select(-c(ea_exp, nea_exp, ea_exp_flip, nea_exp_flip, harmon, harmon1, harmon2,
              rsid_proxy, ea_t2d, nea_t2d, ea_proxy, nea_proxy))
head(clinphewasres)
```

```{r}
ctgwas_sig <- read_tsv("../data/ctgwas_sig.tsv", show_col_types = FALSE)
head(ctgwas_sig)
```

```{r}
ctgwas_sig <- ctgwas_sig %>%
    group_by(trait, profile) %>%
    add_count(name = "nsnps") %>%
    ungroup
head(ctgwas_sig)
```

```{r}
clinphewas <- inner_join(ctgwas_sig, clinphewasres, by = "rsid", relationship = "many-to-many")
head(clinphewas)
```

```{r}
clinphewas <- clinphewas %>%
    filter(abs(eaf_t2d - eaf_exp) < .2)
head(clinphewas)
```

```{r}
clinphewas <- clinphewas %>%
    group_by(trait, rsid, id_exp) %>%
    arrange(desc(reference_snp), desc(eaf_exp), desc(beta_exp/se_exp), .by_group = TRUE) %>%
    slice(1) %>%
    ungroup %>%
    select(-c(r2, reference_snp))
head(clinphewas)
```

```{r}
clinphewas <- clinphewas %>%
    group_by(trait, id_exp, profile) %>%
    add_count(name = "nsnps_exp") %>%
    ungroup
head(clinphewas)
```

```{r}
clinphewas <- clinphewas %>%
    filter(nsnps_exp == nsnps * .8)
head(clinphewas)
```

```{r}
write_tsv(clinphewas, "../data/clinphewas.tsv")
```

```{r}
clinmeta <- clinphewas %>%
    group_by(trait, profile, id_exp, name_exp) %>%
    nest
print(clinmeta)
```

```{r}
clinmeta <- clinmeta %>%
    mutate(
        metares = map(
            data,
            ~tryCatch(
                {
                    PROFILEMETA <- meta::metagen(
                        TE = beta_exp, seTE = se_exp, data = .x,
                        fixed = FALSE, random = TRUE,
                        method.tau = "PM", prediction = FALSE
                    )
                    with(
                        PROFILEMETA,
                        tibble(
                            beta_meta = TE.random, se_meta = seTE.random, pval_meta = pval.random, 
                            conf_low_meta = lower.random, conf_high_meta = upper.random,
                            Qv = Q, Qdf = df.Q, Qp = pval.Q, Tau = tau
                        )
                    )
                },
                error = function(e){ tibble(NULL) }
            )
        )
    )
print(clinmeta)
```

```{r}
clinmeta <- clinmeta %>%
    select(metares) %>%
    unnest(metares) %>%
    ungroup
head(clinmeta)
```

```{r}
write_tsv(clinmeta, "../data/clinmeta.tsv")
```
